refactor(tabs): new hide animation
