+++ src/tabs/demos/zhCN/index.demo-entry.md
@@ -30,6 +30,7 @@ unkeyed-debug.vue
 addable-debug.vue
 animation-debug.vue
 animationx-debug.vue
+none-debug.vue
 ```
 
 ## API
